Move FragmentExtras.kt to :app:ui:base module
