+++ app/ui/base/src/main/java/com/fsck/k9/ui/base/extensions/FragmentExtra.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.ui
+package com.fsck.k9.ui.base.extensions
 
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
